+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - New column for the main channels table: ``PERC_US``: the percentage of funds in the channel that belong to us.
 - New column ``TOTAL_SATS``: the total size of the channel in sats
 - New column ``MIN_HTLC``: the minimum size of an outgoing htlc for that channel in sats
+- New column ``IN_BASE``: the base fee in msats set by the channel's peer
+- New column ``IN_PPM``: the ppm fee set by the channel's peer
 
 ## [3.3.1] 2024-06-29